% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familynblast-data.R
\docType{data}
\name{correct_families}
\alias{correct_families}
\title{All 137 families of flycircuit neurons}
\description{
All 137 families of flycircuit neurons
}
\details{
We still have to get rid of the ones that contains one neuron
}

