% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_loadings_families.R
\name{create_probab_families}
\alias{create_probab_families}
\title{Calculate prior probability for a set of families}
\usage{
create_probab_families(setoffamilies)
}
\arguments{
\item{setoffamilies}{is a list of neurons from the different families.}
}
\description{
The set of families should be of the form: list of families and in each list
the list of neurons in this family. An example is
\code{\link{correct_families}}.
}
\seealso{
\code{\link{correct_families}}
}

