% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familynblast-data.R
\docType{data}
\name{fc_neuron_typec}
\alias{fc_neuron_typec}
\title{All flycircuit neurons with the correct labels we are working on}
\description{
All flycircuit neurons with the correct labels we are working on
}

