% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_loadings_families.R
\name{find_neurons_family}
\alias{find_neurons_family}
\title{Predict the families for a list of neurons}
\usage{
find_neurons_family(listneurons, save = FALSE, path = "",
  computedens = FALSE)
}
\description{
Predict the families for a list of neurons
}
\details{
This will return
}
\seealso{
Other find.family: \code{\link{find_scores_family}},
  \code{\link{neurons_against_fam}}
}

