% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_loadings_families.R
\name{find_percentage_correct_hits}
\alias{find_percentage_correct_hits}
\title{Find the percentage of correct hits within the nth highest scores}
\usage{
find_percentage_correct_hits(listneurons, nb = 3)
}
\arguments{
\item{listneurons}{listneurons is the list of neurons names we want to find the family}

\item{nb}{The number of high scores to consider}
}
\value{
The percentage of correct labelling when we keep the first top score,
the top two scores, top nb scores.
}
\description{
Find the percentage of correct hits within the nth highest scores
}

