% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_loadings_families.R
\name{find_scores_family}
\alias{find_scores_family}
\title{Find scores of list of neurons against all families}
\usage{
find_scores_family(listneurons, computedens = FALSE, zeroscore = -100)
}
\arguments{
\item{zeroscore}{Log score to use when a neuron occupies a supervoxel that is
not overlapped by any family member.}
}
\description{
Find scores of list of neurons against all families
}
\seealso{
Other find.family: \code{\link{find_neurons_family}},
  \code{\link{neurons_against_fam}}
}

