% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familynblast-data.R
\docType{data}
\name{kcs.subfam.training}
\alias{kcs.subfam.test}
\alias{kcs.subfam.training}
\title{Selection of flycircuit Kenyon cells used for subfamily model}
\description{
\code{kcs.subfam.training} consists of 180 neurons in 6
  sub-families, while \code{kcs.subfam.test} are the remaining
  \code{flycircuit} KCs.
}

