% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familynblast-data.R
\docType{data}
\name{number_neurons_fromfam_insv}
\alias{number_neurons_fromfam_insv}
\title{The numbers of neurons crossing each supervoxel, matrix of 7065}
\description{
The numbers of neurons crossing each supervoxel, matrix of 7065

The numbes of neurons crossing each supervoxel, matrix of 7065
}

