% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_loadings_subfamilies.R
\name{prior_prob_nblastscores}
\alias{prior_prob_nblastscores}
\title{Compute the prior probabilities of the nblast scores}
\usage{
prior_prob_nblastscores(subfamilies = kcs.subfam.training, bins = seq(-1, 1,
  0.2))
}
\arguments{
\item{subfamilies}{list of neurons from the training set, neurons type named by flycircuit identifiers}

\item{bins}{bins to cut the nblast scores}
}
\description{
Compute the prior probabilities of the nblast scores
}

