% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_loadings_subfamilies.R
\name{prior_prob_subfam}
\alias{prior_prob_subfam}
\title{Compute the prior probabilities of the subfamilies}
\usage{
prior_prob_subfam(subfamiliesall = kcs)
}
\arguments{
\item{subfamiliesall}{list of neurons neurons type named by flycircuit identifiers, all neurons from the subfamilies}
}
\description{
Compute the prior probabilities of the subfamilies
}

