% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_loadings_subfamilies.R
\name{prior_prob_svscores}
\alias{prior_prob_svscores}
\title{Compute the prior probabilities of the supervoxels}
\usage{
prior_prob_svscores(subfamilies = kcs.subfam.training, computedens = FALSE)
}
\arguments{
\item{subfamilies}{list of neurons from the training set, neurons type named by flycircuit identifiers}

\item{computedens}{if neurons are not from flycircuit dataset}
}
\description{
Compute the prior probabilities of the supervoxels
}

