% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familynblast-data.R
\docType{data}
\name{probability_correct_families}
\alias{probability_correct_families}
\title{The prior probabilities for all the families in correct_families}
\description{
The prior probabilities for all the families in correct_families
}
\examples{
hist(probability_correct_families)
}
\seealso{
\code{\link{correct_families}}
}

