% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familynblast-data.R
\docType{data}
\name{svoxels.fcwb}
\alias{svoxels.fcwb}
\title{Supervoxel label field in FCWB space}
\description{
This was originally supplied by Kristin Branson in the JFRC2 template space
and then bridged to the FCWB space containing the flycircuit neurons. It
contains 7065 non-zero levels defining supervoxels covering nc82 positive
regions in the brain (i.e. neuropil).
}
\seealso{
Other flycircuit-supervoxel-data: \code{\link{svoxels.fcwb.table}},
  \code{\link{voxel_dens_allneurons}}
}

