% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familynblast-data.R
\docType{data}
\name{svoxels.fcwb.table}
\alias{svoxels.fcwb.table}
\title{A table that contains the number of voxels in the supervoxels}
\description{
A table that contains the number of voxels in the supervoxels
}

