% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_loadings_families.R
\name{voxel_dens}
\alias{voxel_dens}
\title{Computes the number of points from a setofneurons in each supervoxel}
\usage{
voxel_dens(setofneurons, svoxels = familynblast::svoxels.fcwb,
  svoxels.table = familynblast::svoxels.fcwb.table)
}
\arguments{
\item{setofneurons}{\code{\link[nat]{neuronlist}} object containing set}

\item{svoxels}{A \code{\link[nat]{im3d}} object containing defining a set of
supervoxels. Defaults to \code{\link{svoxels.fcwb}}.}

\item{svoxels.table}{A \code{\link{table}} indicating how many pixels are
present in each supervoxel. Defaults to \code{\link{svoxels.fcwb.table}}.}
}
\value{
A \code{matrix} with length(setofneurons) rows and
  \code{length(svoxels.table)} columns.
}
\description{
Computes the number of points from a setofneurons in each supervoxel
}
\examples{
# convert a sample set of neurons to supervoxel representation
kcdens=voxel_dens(nat::kcs20)
}

