% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familynblast-data.R
\docType{data}
\name{voxel_dens_allneurons}
\alias{voxel_dens_allneurons}
\title{Precalculated super voxel density representation of flycircuit neurons}
\description{
This object is a sparse \code{\link[Matrix]{Matrix}} since it has many zeros.
It has 7066 columns (all supervoxels including catch-all 0 level) and 16129
rows for each flycircuit neuron. The rows are named with the flycircuit
\code{gene_name} - see \code{\link[flycircuit]{fc_gene_name}}. CHECK
}
\seealso{
Other flycircuit-supervoxel-data: \code{\link{svoxels.fcwb.table}},
  \code{\link{svoxels.fcwb}}
}

