% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_loadings_families.R
\name{create_probab_sv_knowing_fam}
\alias{create_probab_sv_knowing_fam}
\title{Create probability matrix with supervoxels as rows and families as columns for a setoffamilies}
\usage{
create_probab_sv_knowing_fam(setoffamilies, computedens = FALSE)
}
\arguments{
\item{computedens}{}
}
\description{
Create probability matrix with supervoxels as rows and families as columns for a setoffamilies
}

