% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_loadings_families.R
\name{find_percentage_correct_hits}
\alias{find_percentage_correct_hits}
\title{Find the percentage of correct hits within the nth highest scores}
\usage{
find_percentage_correct_hits(listneurons, nb = 3)
}
\arguments{
\item{listneurons}{FIXME}

\item{nb}{The number of high scores to consider}
}
\value{
FIXME
}
\description{
Find the percentage of correct hits within the nth highest scores
}

