% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_loadings_families.R
\name{neurons_against_fam}
\alias{neurons_against_fam}
\title{Computing the score of one neuron against a particular family, given the
index of the name of the family in correct_families}
\usage{
neurons_against_fam(listneurons, familyind, computedens = FALSE,
  zeroscore = -100)
}
\arguments{
\item{computedens}{Whether or not to compute desnsities (which would be
necessary for non-flycircuit neurons)}

\item{zeroscore}{Log score to use when a neuron occupies a supervoxel that is
not overlapped by any family member.}
}
\description{
Computing the score of one neuron against a particular family, given the
index of the name of the family in correct_families
}
\seealso{
Other find.family: \code{\link{find_neurons_family}},
  \code{\link{find_scores_family}}
}

