% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familynblast-data.R
\docType{data}
\name{number_neurons_fromfam_insv}
\alias{number_neurons_fromfam_insv}
\title{The numbers of neurons crossing each supervoxel}
\description{
Matrix of 7065 rows (supervoxels) and 137 columns (families)
}
\seealso{
Other flycircuit-family-data: \code{\link{correct_families}},
  \code{\link{fc_neuron_typec}},
  \code{\link{probability_correct_families}},
  \code{\link{probability_sv_knowing_family}}
}

