% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familynblast-data.R
\docType{data}
\name{probability_sv_knowing_family}
\alias{probability_sv_knowing_family}
\title{The probabilities of all the supervoxels in each family, matrix of nrow = 7065}
\description{
The probabilities of all the supervoxels in each family, matrix of nrow = 7065
}

