% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familynblast-data.R
\docType{data}
\name{probability_sv_knowing_subfamily}
\alias{probabilities_nblastscores_kcs}
\alias{probability_subfamily_kcs}
\alias{probability_sv_knowing_subfamily}
\title{Probability matrices for flycircuit Kenyon cell subfamily model}
\description{
\code{probability_sv_knowing_subfamily} is a 7065 x 6 matrix
  describing the probability of each supervoxel being crossed by one of the
  neurons in the \code{\link{kcs.subfam.training}} training set. Computed by
  \code{\link{prior_prob_svscores}}.

  \code{probability_subfamily_kcs} contains the prior probability for each of
  the 6 subfamilies in the complete flycircuit dataset. Computed by
  \code{\link{prior_prob_subfam}}.

\code{probabilities_nblastscores_kcs} contains the probabilities
  for NBLAST scores to fall into a certain bin (11 bins in range -1 to +1 )
  for neurons in the 6 different KC families. It is a 3D array with
  dimensions. Computed by \code{\link{prior_prob_nblastscores}}.
}
\examples{
# dimensions of NBLAST probabilities by subfamily
dim(probabilities_nblastscores_kcs)
dimnames(probabilities_nblastscores_kcs)
}
\seealso{
\code{\link{prior_prob_svscores}}, \code{\link{prior_prob_subfam}},
  \code{\link{prior_prob_nblastscores}}
}

